feat: use single-file components
