+++ apps/this-is-learning/src/app/app.module.ts
@@ -1,13 +1,12 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { ScullyLibModule } from '@scullyio/ng-lib';
 
 import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
+import { AppComponent, AppScam } from './app.component';
+import { ScullyLibModule } from '@scullyio/ng-lib';
 
 @NgModule({
   bootstrap: [AppComponent],
-  declarations: [AppComponent],
-  imports: [BrowserModule, AppRoutingModule, ScullyLibModule],
+  imports: [BrowserModule, AppScam, AppRoutingModule, ScullyLibModule],
 })
 export class AppModule {}
