--- apps/this-is-learning/src/app/hello/hello.module.ts
@@ -1,10 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-
-import { HelloComponent } from './hello.component';
-
-@NgModule({
-  declarations: [HelloComponent],
-  imports: [CommonModule],
-})
-export class HelloModule {}
